EDM-427: Separate out image pkg
