+++ internal/agent/device/spec/spec_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/flightctl/flightctl/api/v1alpha1"
 	"github.com/flightctl/flightctl/internal/agent/client"
 	"github.com/flightctl/flightctl/internal/agent/device/fileio"
+	"github.com/flightctl/flightctl/internal/agent/device/image"
 	"github.com/flightctl/flightctl/internal/container"
 	"github.com/flightctl/flightctl/pkg/log"
 	"github.com/stretchr/testify/require"
@@ -513,7 +514,7 @@ func TestCheckOsReconciliation(t *testing.T) {
 
 		bootedOSImage, desiredImageIsBooted, err := s.CheckOsReconciliation(ctx)
 		require.NoError(err)
-		require.Equal(&Image{Base: bootedImage}, bootedOSImage)
+		require.Equal(&image.Image{Base: bootedImage}, bootedOSImage)
 		require.Equal(false, desiredImageIsBooted)
 	})
 
@@ -531,24 +532,24 @@ func TestCheckOsReconciliation(t *testing.T) {
 
 		bootedOSImage, desiredImageIsBooted, err := s.CheckOsReconciliation(ctx)
 		require.NoError(err)
-		require.Equal(&Image{Base: "flightctl-device", Tag: "v1"}, bootedOSImage)
+		require.Equal(&image.Image{Base: "flightctl-device", Tag: "v1"}, bootedOSImage)
 		require.Equal(false, desiredImageIsBooted)
 	})
 
 	t.Run("booted image and desired image are the same", func(t *testing.T) {
-		image := "flightctl-device:v2"
+		testImage := "flightctl-device:v2"
 
 		bootcStatus := &container.BootcHost{}
-		bootcStatus.Status.Booted.Image.Image.Image = image
+		bootcStatus.Status.Booted.Image.Image.Image = testImage
 		mockBootcClient.EXPECT().Status(ctx).Return(bootcStatus, nil)
 
-		desiredSpec, err := createTestSpec(image)
+		desiredSpec, err := createTestSpec(testImage)
 		require.NoError(err)
 		mockReadWriter.EXPECT().ReadFile(desiredPath).Return(desiredSpec, nil)
 
 		bootedOSImage, desiredImageIsBooted, err := s.CheckOsReconciliation(ctx)
 		require.NoError(err)
-		require.Equal(&Image{Base: "flightctl-device", Tag: "v2"}, bootedOSImage)
+		require.Equal(&image.Image{Base: "flightctl-device", Tag: "v2"}, bootedOSImage)
 		require.Equal(true, desiredImageIsBooted)
 	})
 }
@@ -1128,163 +1129,6 @@ func Test_getRenderedVersion(t *testing.T) {
 
 			require.NoError(err)
 			require.Equal(testCase.expectedReturnValue, renderedVersion)
-		})
-	}
-}
-
-func Test_areImagesEquivalent(t *testing.T) {
-	require := require.New(t)
-
-	digest := "sha256:6cf77c2a98dd4df274d14834fab9424b6e96ef3ed3f49f792b27c163763f52b5"
-	digestTwo := "sha256:bd1b50c5a1df1bcb701e3556075a890c4e4a87765f985ee3a4b87df91db98c4d"
-	testCases := []struct {
-		name           string
-		imageOne       *Image
-		imageTwo       *Image
-		expectedResult bool
-	}{
-		{
-			name:           "both are nil",
-			imageOne:       nil,
-			imageTwo:       nil,
-			expectedResult: true,
-		},
-		{
-			name:           "one is defined and the other nil",
-			imageOne:       nil,
-			imageTwo:       &Image{},
-			expectedResult: false,
-		},
-		{
-			name: "image digests are equal",
-			imageOne: &Image{
-				Digest: digest,
-			},
-			imageTwo: &Image{
-				Digest: digest,
-			},
-			expectedResult: true,
-		},
-		{
-			name: "image digests are not equal",
-			imageOne: &Image{
-				Digest: digest,
-			},
-			imageTwo: &Image{
-				Digest: digestTwo,
-			},
-			expectedResult: false,
-		},
-		{
-			name: "image bases match",
-			imageOne: &Image{
-				Base: "flightct-device",
-			},
-			imageTwo: &Image{
-				Base: "flightct-device",
-			},
-			expectedResult: true,
-		},
-		{
-			name: "image bases match when one image has a digest defined",
-			imageOne: &Image{
-				Base:   "flightct-device",
-				Digest: digest,
-			},
-			imageTwo: &Image{
-				Base: "flightct-device",
-			},
-			expectedResult: true,
-		},
-		{
-			name: "image bases are different",
-			imageOne: &Image{
-				Base: "flightct-device",
-			},
-			imageTwo: &Image{
-				Base: "device-os",
-			},
-			expectedResult: false,
-		},
-		{
-			name: "image bases are different but digests are identical",
-			imageOne: &Image{
-				Base:   "flightct-device",
-				Digest: digest,
-			},
-			imageTwo: &Image{
-				Base:   "device-os",
-				Digest: digest,
-			},
-			expectedResult: true,
-		},
-		{
-			name: "image bases match and one has a tag",
-			imageOne: &Image{
-				Base: "flightct-device",
-				Tag:  "v1",
-			},
-			imageTwo: &Image{
-				Base: "flightct-device",
-			},
-			expectedResult: false,
-		},
-		{
-			name: "image bases match and tags match",
-			imageOne: &Image{
-				Base: "flightct-device",
-				Tag:  "v2",
-			},
-			imageTwo: &Image{
-				Base: "flightct-device",
-				Tag:  "v2",
-			},
-			expectedResult: true,
-		},
-		{
-			name: "image bases match and tags are different",
-			imageOne: &Image{
-				Base: "flightct-device",
-				Tag:  "v2",
-			},
-			imageTwo: &Image{
-				Base: "flightct-device",
-				Tag:  "v9",
-			},
-			expectedResult: false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			result := AreImagesEquivalent(testCase.imageOne, testCase.imageTwo)
-			require.Equal(testCase.expectedResult, result)
-		})
-	}
-}
-
-func Test_parseImage(t *testing.T) {
-	require := require.New(t)
-	testCases := []struct {
-		name           string
-		image          string
-		expectedResult *Image
-	}{
-		{
-			name:  "image with a tag and digest",
-			image: "flightctl-device:v3@sha256:123abc",
-			expectedResult: &Image{
-				Base:   "flightctl-device",
-				Tag:    "v3",
-				Digest: "sha256:123abc",
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			result := parseImage(testCase.image)
-			require.Equal(testCase.expectedResult, result)
 		})
 	}
 }
